% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivarrugplot.r
\name{bivarrugplot}
\alias{bivarrugplot}
\title{Draw a scatterplot with marginal rugs}
\usage{
bivarrugplot(x, y, data)
}
\arguments{
\item{x}{The name of the x column, as a string.}

\item{y}{The name of the y column, as a string.}

\item{data}{A data frame.}
}
\value{
Prints the plot and returns the ggplot object invisibly.
}
\description{
Draws a scatterplot and adds marks along each axis at the observed values.
}
\examples{
a <- runif(1000, min = 0, max = 1)
b <- a + rnorm(1000, mean = 0, sd = 1)
data <- data.frame(a, b)
bivarrugplot("a", "b", data)
}
\author{
Charles Crabtree \email{charles.crabtree@monash.edu}
}
