% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.r
\name{clear}
\alias{clear}
\title{Clear the console display}
\usage{
clear(...)
}
\arguments{
\item{...}{An unused argument.}
}
\value{
No return value (called for side effect of printing a form feed
 character to the console, which visually clears the R terminal). Returns
 \code{NULL} invisibly.
}
\description{
Prints a form-feed character. Whether this clears the display depends on your console.
}
\examples{
clear()
}
\author{
Charles Crabtree \email{charles.crabtree@monash.edu}
}
