% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcor.r
\name{groupcor}
\alias{groupcor}
\title{Calculates the correlation between measures within groups}
\usage{
groupcor(x, y, unit, plotdata, stat)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{unit}{A vector that contians unit identifiers.}

\item{plotdata}{A data frame.}

\item{stat}{The type of correlation coefficient to be calculated. One of "pearson", "kendall", or "spearman".}
}
\value{
A data frame with unit identifiers and calculated correlation coefficients.
}
\examples{
\dontrun{
x <- runif(1000, min = 0, max = 1)
y <- x + rnorm(1000, mean = 0, sd =1)
z <- rep(c(1:10), times = 100)
data <- data.frame(x, y, z)
groupcor(x, y, z, data)
print(dat)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
\seealso{
\code{\link{bivarplots}}, \code{\link{groupcorplots}}
}

