% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcorplot.r
\name{groupcorplot}
\alias{groupcorplot}
\title{Plots the correlation coefficient between measures across units}
\usage{
groupcorplot(x, y, unit, plotdata, stat)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{unit}{A vector that contians unit identifiers.}

\item{plotdata}{A data frame.}

\item{stat}{The type of correlation coefficient to be calculated. One of "pearson", "kendall", or "spearman".}
}
\value{
A figure that plots the correlation coefficient between measures across units.
}
\examples{
\dontrun{
data$x <- runif(1000, min = 0, max = 1)
data$y <- x + rnorm(1000, mean = 0, sd =1)
data$z <- rep(c(1:10), times = 100)
groupcorplot(x, y, z, data, "pearson")
print(dat)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
\seealso{
\code{\link{bivarplots}}, \code{\link{groupcor}}
}

