% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthunique.r
\name{length_unique}
\alias{length_unique}
\title{Calculates the number of unique values in a vector}
\usage{
length_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The number of unique values in a vector.
}
\description{
Calculates the number of unique values in a vector.
}
\examples{
\dontrun{
x <- rep(c(1:10), 10)
length_unique(x)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}

