% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthunique.r
\name{lengthunique}
\alias{lengthunique}
\title{Count distinct non-missing values}
\usage{
lengthunique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The number of distinct values after removing missing values.
}
\description{
Count distinct non-missing values.
}
\examples{
x <- rep(c(1:10), 10)
lengthunique(x)
}
\author{
Charles Crabtree \email{charles.crabtree@monash.edu}
}
