% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.r
\name{save.package.list}
\alias{save.package.list}
\title{Prints the list of installed packages to a text file}
\usage{
save.package.list(wd = NULL, file = NULL)
}
\arguments{
\item{wd}{The directory to which the text file will be exported. The default is the working directory.}

\item{file}{The name of the exported text file. The default is \code{installed_r_packages.txt}.}
}
\value{
A text file that contains a list of all installed packages.
}
\description{
Print the list of installed packages to a text file.
}
\examples{
\dontrun{
save.package.list(wd = "~/", file = "packages.txt")
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}

