% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.r
\name{clear}
\alias{clear}
\title{(Effectively) clears R terminal}
\usage{
clear(...)
}
\arguments{
\item{...}{An unused argument.}
}
\description{
Effectively clears the R terminal by filling it with whitespace.
}
\examples{
\dontrun{
clear()
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
