% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcor.r
\name{groupcor}
\alias{groupcor}
\title{Calculates the correlation between measures within units/groups}
\usage{
groupcor(x, y, unit, data, stat)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{unit}{A vector that contains unit/group identifiers.}

\item{data}{A data frame.}

\item{stat}{The type of correlation coefficient to be calculated. One of "pearson", "kendall", or "spearman".}
}
\value{
A tibble data frame with unit identifiers and calculated correlation coefficients.
}
\examples{
\dontrun{
a <- runif(1000, min = 0, max = 1)
b <- x + rnorm(1000, mean = 0, sd =1)
c <- rep(c(1:10), times = 100)
data <- data.frame(a, b, c)
groupcor("a", "b", "c", data, "pearson")
print(dat)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
\seealso{
\code{\link{bivarplots}}, \code{\link{groupcorplots}}
}

