% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthunique.r
\name{lengthunique}
\alias{lengthunique}
\title{Calculates the number of unique values in a vector}
\usage{
lengthunique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The number of unique values in a vector.
}
\description{
Calculates the number of unique values in a vector.
}
\examples{
x <- rep(c(1:10), 10)
lengthunique(x)
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
