% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefacnum.r
\name{makefacnum}
\alias{makefacnum}
\title{Converts factor vectors to numeric vectors}
\usage{
makefacnum(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A numeric vector.
}
\description{
Converts factor vectors to numeric vectors.
}
\examples{
\dontrun{
x <- c("x", "y", "z")
x <- as.factor(x)
x
x <- makefacnum(x)
x
is.numeric(x)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
