% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplots.r
\name{violinplots}
\alias{violinplots}
\title{Creates violin plots for a relationship between two measures across units}
\usage{
violinplots(x, y, unit, data)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{unit}{A vector that contains unit/group identifiers.}

\item{data}{A data frame.}

\item{stat}{The type of correlation coefficient to be calculated. One of "pearson", "kendall", or "spearman".}
}
\value{
A tibble data frame with unit identifiers and calculated correlation coefficients.
}
\examples{
\dontrun{
a <- runif(1000, min = 0, max = 1)
b <- x + rnorm(1000, mean = 0, sd =1)
c <- rep(c(1:10), times = 100)
data <- data.frame(a, b, c)
violinplots("a", "b", "c", data)
print(dat)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}

